+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.3
+* Allow to provide custom getEvent method to tooltip service
+
+## 2.0.2
+* Convert tooltiputils to es2015 modules
+
 ## 1.0.0
 * Removed `typings`
 * Unified dependencies versions